+++ packages/scripts/spreadsheet_helpers/bum_tones/INSTRUCTIONS.md
@@ -1,0 +1,8 @@
+Before running scripts, change in spreadsheet: L H LH and HL characters for unicode accents:
+
+- \u030C = ˇ (hacek)
+- \u0300 = ` (Grave accent)
+- \u0301 = ´ (Acute accent)
+- \u0302 = ˆ (Circumflex accent)
+
+check test_tones.ts example
